bug fixing
fixing some bugs, issues with key event implementation

+++ src/Point.java
@@ -4,10 +4,10 @@
 
 public class Point {
 
-    public int x;
-    public int y;
+    public float x;
+    public float y;
 
-    public Point(int x, int y) {
+    public Point(float x, float y) {
         this.x = x;
         this.y = y;
     }
